[#174] feat: profile Img 가져오기
