+++ src/components/organisms/RegisterForm/index.tsx
@@ -5,24 +5,50 @@ import { useFormContext } from 'react-hook-form';
 import { UserInputRegisterProps } from '@/types/user';
 import ConfirmButton from '@/components/atoms/Button/ConfirmButton';
 import { useState } from 'react';
-
-interface RegisterFormProp {
-  onSubmit: () => void;
-}
+import { postRegister } from '@/apis/user/register';
+import { useRouter } from 'next/navigation';
+import useImageUpload from '@/hooks/useImageUpload';
 
 //TODO: 디자인 입히기
-function RegisterForm({ onSubmit }: RegisterFormProp) {
+function RegisterForm(methods: any) {
   const {
     register,
     formState: { errors },
   } = useFormContext<UserInputRegisterProps>();
 
+  const router = useRouter();
+  const { handleUploadImageToS3 } = useImageUpload();
+  const [image, setImage] = useState('');
+
+  const onSubmit = async (data: UserInputRegisterProps) => {
+    if (image) {
+      const imageUrl = await handleUploadImageToS3();
+      if (imageUrl) {
+        data.profileUrl = imageUrl;
+      }
+    }
+
+    try {
+      await postRegister(data);
+      router.push('/');
+    } catch (e) {
+      throw Error('회원가입에 실패하였습니다');
+    }
+  };
+
+  const handleImageChange = (image: string) => {
+    setImage(image);
+  };
+
   const [disabled, setDisabled] = useState(true);
 
   return (
-    <div className={styles.wrapper}>
+    <form className={styles.wrapper} onSubmit={methods.handleSubmit(onSubmit)}>
       <Text onBoarding="registerExplain">프로필을 설정해주세요.</Text>
-      <ProfileImgSetting register={register('profileUrl')} />
+      <ProfileImgSetting
+        onImageChange={handleImageChange}
+        {...register('profileUrl')}
+      />
 
       <div className={styles.nicknameWrapper}>
         <input placeholder="닉네임" {...register('nickname')} />
@@ -30,11 +56,9 @@ function RegisterForm({ onSubmit }: RegisterFormProp) {
 
       <h1>{errors.nickname?.message}</h1>
 
-      <button type="submit" onClick={onSubmit}>
-        완료
-      </button>
+      <button type="submit">완료</button>
       {/* <ConfirmButton title="완료" type="submit" /> */}
-    </div>
+    </form>
   );
 }
